modularized GameCards
